Rename card scheme icons to kebab-case
